Started splitting menu class into their own files.
This effort is especially made to ease later maintenance of the inventory part.
